requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
argcomplete==3.6.2
attrs==25.3.0
audioop-lts==0.2.1
build==1.2.2.post1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click>=8.1
click-option-group==0.5.7
colorama==0.4.6
commitizen==4.8.2
decli==0.6.3
Deprecated==1.2.18
dotty-dict==1.3.1
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
packaging==25.0
pip-tools==7.4.1
prompt_toolkit==3.0.51
propcache==0.3.1
py-cord==2.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyNaCl==1.5.0
pyproject_hooks==1.2.0
python-dotenv==1.1.0
python-gitlab==5.6.0
python-semantic-release==10.0.2
PyYAML==6.0.2
questionary==2.1.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
setuptools==80.9.0
shellingham==1.5.4
smmap==5.0.2
termcolor==2.5.0
tomlkit==0.13.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
